:recycle:refactoring redirectview
